+++ crayg/image/ImageWriter.h
@@ -8,7 +8,6 @@
 #include <string>
 #include "Image.h"
 
-// todo add virtual destructor
 class ImageWriter{
 public:
     virtual void writeImage(const Image &image, std::string image_name) = 0;