+++ standalone/main.cpp
@@ -6,6 +6,7 @@
 #include <image/ImageWriterFactory.h>
 #include <utils/ToStringHelper.h>
 #include <intersectors/SceneIntersector.h>
+#include <Renderer.h>
 #include "scene/Camera.h"
 #include "PineHoleCameraModel.h"
 #include "image/ImageIterators.h"
@@ -22,7 +23,6 @@ int main(int argc, char *argv[])
     Image myImage(800,600);
 
     Camera camera(Vector3f(0,0,10), Vector3f(0,1,0), Vector3f(0,0,-1), 6.0f*180.0f/3.141f);
-    PineHoleCameraModel cameraModel(camera,myImage.getWidth(),myImage.getHeight());
 
     Scene scene;
     SceneIntersector sceneIntersector(scene);
@@ -35,39 +35,11 @@ int main(int argc, char *argv[])
     }
     std::cout << "Sphere generation done." << std::endl;
 
-    std::cout << ToStringHelper("Pixel")
-            .addMember("x", 6)
-            .addMember("y", 7)
-            .finish() << std::endl;
-
-
-    for(int x=0; x<myImage.getWidth();x++){
-        for(int y=0; y<myImage.getHeight();y++){
-            Ray ray = cameraModel.createPrimaryRay(x,y);
-
-            if(sceneIntersector.intersects(ray)){
-                myImage.setValue(x,y,1,1,1);
-            }
-        }
-    }
+    Renderer renderer(scene, camera, myImage);
+    renderer.renderScene();
 
     std::unique_ptr<ImageWriter> imageWriter(createImageWriter(ImageWriterType::BMP));
     imageWriter->writeImage(myImage, "img.bmp");
-
-    std::cout << "write" << std::endl;
-
-    JsonSerializer serializer("test.json");
-    SceneWriter sceneWriter(scene, serializer);
-    sceneWriter.write();
-
-    std::cout << "read" << std::endl;
-    Scene newScene;
-    SceneReader sceneReader(newScene);
-    sceneReader.read("test.json");
-
-    std::cout << "done" << std::endl;
-
-    std::cout << newScene.objects.size() << std::endl;
 
     return 0;
 }