+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SceneWriter: added tests
 - JsonSerializer: Created abstract class SerializerImplementation for better abstraction
 - ImageWriter: added virtual destructor
+- Added Renderer Class
 
 ## [0.1.0]
 ### Added